Refactor inference rules.
Rules now define both a 'params' and 'infer' method. The former will
construct the necessary rule parameters for inference given the current
proof context, if defined for the Rule; the latter will construct a new
Proof from the given RuleArgs if and only if they satisfy the inference
requirements.

Connectives and quantifiers also now have apply/unapply defined for their
relevant Sentence and Term operands. This means that the operator types
can now be used directly in pattern matching instead of via matching on
the Sentence case types.